Begin restructuring plugins
