--- src/core/plugins/adb.spec.js
@@ -1,3 +1,0 @@
-const adb = require("./adb.js");
-
-it("should be a singleton", () => expect(adb).toEqual(require("./adb.js")));
