+++ src/core/plugins/adb/plugin.spec.js
@@ -1,0 +1,3 @@
+const adb = require("./plugin.js");
+
+it("should be a singleton", () => expect(adb).toEqual(require("./plugin.js")));
