+++ src/core/plugins/core/plugin.js
@@ -20,10 +20,10 @@
 const fs = require("fs-extra");
 const path = require("path");
 const { download, checkFile } = require("progressive-downloader");
-const mainEvent = require("../../lib/mainEvent.js");
-const { unpack } = require("../../lib/asarLibs.js");
-const log = require("../../lib/log.js");
-const { path: cachePath } = require("../../lib/cache.js");
+const mainEvent = require("../../../lib/mainEvent.js");
+const { unpack } = require("../../../lib/asarLibs.js");
+const log = require("../../../lib/log.js");
+const { path: cachePath } = require("../../../lib/cache.js");
 
 /**
  * core actions plugin
