--- src/core/plugins/heimdall.spec.js
@@ -1,4 +1,0 @@
-const heimdall = require("./heimdall.js");
-
-it("should be a singleton", () =>
-  expect(heimdall).toEqual(require("./heimdall.js")));
