+++ src/core/plugins/heimdall/plugin.spec.js
@@ -1,0 +1,4 @@
+const heimdall = require("./plugin.js");
+
+it("should be a singleton", () =>
+  expect(heimdall).toEqual(require("./plugin.js")));
