--- src/core/plugins/systemimage.spec.js
@@ -1,4 +1,0 @@
-const systemimage = require("./systemimage.js");
-
-it("should be a singleton", () =>
-  expect(systemimage).toEqual(require("./systemimage.js")));
