+++ src/core/plugins/systemimage/plugin.spec.js
@@ -1,0 +1,4 @@
+const systemimage = require("./plugin.js");
+
+it("should be a singleton", () =>
+  expect(systemimage).toEqual(require("./plugin.js")));
